+++ app/src/main/AndroidManifest.xml
@@ -319,7 +319,7 @@
     </receiver>
 
     <service android:name=".notification.PullingContentService"/>
-    <service android:name=".sync.SyncService"/>
+    <service android:name=".sync.alarm.AlarmSyncService"/>
 
     <receiver android:name=".notification.NotificationReceiver">
       <intent-filter>
