+++ app/src/main/java/cm/aptoide/pt/sync/alarm/AlarmSyncService.java
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.sync;
+package cm.aptoide.pt.sync.alarm;
 
 import android.app.Service;
 import android.content.Intent;
@@ -6,8 +6,10 @@ import android.os.IBinder;
 import android.support.annotation.Nullable;
 import cm.aptoide.pt.AptoideApplication;
 import cm.aptoide.pt.crashreports.CrashReport;
+import cm.aptoide.pt.sync.Sync;
+import cm.aptoide.pt.sync.SyncScheduler;
 
-public class SyncService extends Service {
+public class AlarmSyncService extends Service {
 
   private SyncScheduler scheduler;
   private SyncStorage storage;
@@ -15,19 +17,19 @@ public class SyncService extends Service {
 
   @Override public void onCreate() {
     super.onCreate();
-    scheduler = ((AptoideApplication) getApplicationContext()).getSyncScheduler();
+    scheduler = ((AptoideApplication) getApplicationContext()).getAlarmSyncScheduler();
     storage = ((AptoideApplication) getApplicationContext()).getSyncStorage();
     crashReport = CrashReport.getInstance();
   }
 
   @Override public int onStartCommand(Intent intent, int flags, int startId) {
 
-    if (intent != null && SyncScheduler.ACTION_SYNC.equals(intent.getAction())) {
+    if (intent != null && AlarmSyncScheduler.ACTION_SYNC.equals(intent.getAction())) {
 
       final String syncId = intent.getData()
           .getFragment();
       final Sync sync = storage.get(syncId);
-      final boolean reschedule = intent.getBooleanExtra(SyncScheduler.EXTRA_RESCHEDULE, false);
+      final boolean reschedule = intent.getBooleanExtra(AlarmSyncScheduler.EXTRA_RESCHEDULE, false);
 
       if (sync != null) {
         sync.execute()
